+++ app/src/main/java/org/android/bbangzip/data/source/remote/service/di/ServiceModule.kt
@@ -6,7 +6,7 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import org.android.bbangzip.data.source.remote.service.DummyService
 import org.android.bbangzip.data.source.remote.service.TimerService
-import org.android.bbangzip.data.source.remote.util.qualifier.BbangZip
+import org.android.bbangzip.data.network.auth.qualifier.BbangZip
 import retrofit2.Retrofit
 import javax.inject.Singleton
